+++ src/main/java/nl/enjarai/showmeyourskin/gui/widget/ArmorConfigWindow.java
@@ -2,7 +2,6 @@ package nl.enjarai.showmeyourskin.gui.widget;
 
 import com.google.common.collect.Lists;
 import com.mojang.blaze3d.systems.RenderSystem;
-import it.unimi.dsi.fastutil.booleans.BooleanConsumer;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.AbstractParentElement;
 import net.minecraft.client.gui.Drawable;
@@ -11,15 +10,9 @@ import net.minecraft.client.gui.Selectable;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.gui.screen.ingame.InventoryScreen;
 import net.minecraft.client.gui.screen.narration.NarrationMessageBuilder;
-import net.minecraft.client.gui.widget.ButtonWidget;
 import net.minecraft.client.gui.widget.ClickableWidget;
-import net.minecraft.client.gui.widget.SliderWidget;
-import net.minecraft.client.gui.widget.TexturedButtonWidget;
-import net.minecraft.client.render.GameRenderer;
 import net.minecraft.client.sound.PositionedSoundInstance;
-import net.minecraft.client.sound.SoundManager;
 import net.minecraft.client.util.math.MatrixStack;
-import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.sound.SoundEvents;
 import net.minecraft.text.Text;
@@ -27,21 +20,23 @@ import net.minecraft.util.Identifier;
 import nl.enjarai.showmeyourskin.ShowMeYourSkin;
 import nl.enjarai.showmeyourskin.config.ArmorConfig;
 import nl.enjarai.showmeyourskin.config.ModConfig;
-import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
 
 public class ArmorConfigWindow extends AbstractParentElement implements Drawable, Element, Selectable {
     public static final Identifier TEXTURE = ShowMeYourSkin.id("textures/gui/armor_screen.png");
+    public static final Identifier BACKGROUND_TEXTURE = ShowMeYourSkin.id("textures/gui/armor_screen_background.png");
     public static final Identifier OVERLAY_TEXTURE = ShowMeYourSkin.id("textures/gui/armor_screen_disabled.png");
     private static final int TEXT_COLOR = 0x303030;
-    private static final Text GLINT_TOOLTIP = Text.translatable("gui.showmeyourskin.armorScreen.glintTooltip");
     private static final Text COMBAT_TOOLTIP = Text.translatable("gui.showmeyourskin.armorScreen.combatTooltip");
     private static final Text NAME_TAG_TOOLTIP = Text.translatable("gui.showmeyourskin.armorScreen.nameTagTooltip");
     private static final Text SHOW_ELYTRA_TOOLTIP = Text.translatable("gui.showmeyourskin.armorScreen.showElytraTooltip");
     private static final Text SHIELD_GLINT_TOOLTIP = Text.translatable("gui.showmeyourskin.armorScreen.shieldGlintTooltip");
 
     private final List<ClickableWidget> buttons = Lists.newArrayList();
+    private final List<Element> children = Lists.newArrayList();
+    private final List<SliderSetTab> sliderSetTabs = Lists.newArrayList();
+    private SliderSetTab selectedSliderSetTab;
     private final Screen parent;
     public int x;
     public int y;
@@ -58,28 +53,28 @@ public class ArmorConfigWindow extends AbstractParentElement implements Drawable
         this.player = player;
         this.armorConfig = armorConfig;
 
-        buttons.add(getTransparencySlider(EquipmentSlot.HEAD, 14, 11,
-                "gui.showmeyourskin.armorScreen.headSlider"));
-        buttons.add(getTransparencySlider(EquipmentSlot.CHEST, 14, 35,
-                "gui.showmeyourskin.armorScreen.chestSlider"));
-        buttons.add(getTransparencySlider(EquipmentSlot.LEGS, 14, 59,
-                "gui.showmeyourskin.armorScreen.legsSlider"));
-        buttons.add(getTransparencySlider(EquipmentSlot.FEET, 14, 83,
-                "gui.showmeyourskin.armorScreen.feetSlider"));
+        sliderSetTabs.add(new SliderSetTab(getWindowLeft() - 25, getWindowTop() + 12, 0, 0,
+                new SliderSet(this, getWindowLeft(), getWindowTop(),
+                        "gui.showmeyourskin.armorScreen.piece", armorConfig.getPieces())));
+        sliderSetTabs.add(new SliderSetTab(getWindowLeft() - 25, getWindowTop() + 42, 0, 0,
+                new SliderSet(this, getWindowLeft(), getWindowTop(),
+                        "gui.showmeyourskin.armorScreen.trim", armorConfig.getTrims())));
+        sliderSetTabs.add(new SliderSetTab(getWindowLeft() - 25, getWindowTop() + 72, 0, 0,
+                new SliderSet(this, getWindowLeft(), getWindowTop(),
+                        "gui.showmeyourskin.armorScreen.glint", armorConfig.getGlints())));
 
-        buttons.add(getGlintButton(EquipmentSlot.HEAD, 94, 11));
-        buttons.add(getGlintButton(EquipmentSlot.CHEST, 94, 35));
-        buttons.add(getGlintButton(EquipmentSlot.LEGS, 94, 59));
-        buttons.add(getGlintButton(EquipmentSlot.FEET, 94, 83));
+        selectedSliderSetTab = sliderSetTabs.get(0);
 
-        buttons.add(new ToggleButtonWidget(parent, getWindowLeft() + 14, getWindowTop() + 107, 40, 38,
+        buttons.add(new ToggleButtonWidget(parent, getWindowLeft() + 14, getWindowTop() + 115, 40, 38,
                 TEXTURE, armorConfig.showInCombat, b -> armorConfig.showInCombat = b, COMBAT_TOOLTIP));
-        buttons.add(new ToggleButtonWidget(parent, getWindowLeft() + 38, getWindowTop() + 107, 80, 38,
+        buttons.add(new ToggleButtonWidget(parent, getWindowLeft() + 40, getWindowTop() + 115, 80, 38,
                 TEXTURE, armorConfig.showNameTag, b -> armorConfig.showNameTag = b, NAME_TAG_TOOLTIP));
-        buttons.add(new ToggleButtonWidget(parent, getWindowLeft() + 62, getWindowTop() + 107, 120, 38,
+        buttons.add(new ToggleButtonWidget(parent, getWindowLeft() + 66, getWindowTop() + 115, 120, 38,
                 TEXTURE, armorConfig.showElytra, b -> armorConfig.showElytra = b, SHOW_ELYTRA_TOOLTIP));
-        buttons.add(new ToggleButtonWidget(parent, getWindowLeft() + 94, getWindowTop() + 107, 160, 38,
+        buttons.add(new ToggleButtonWidget(parent, getWindowLeft() + 92, getWindowTop() + 115, 160, 38,
                 TEXTURE, armorConfig.showShieldGlint, b -> armorConfig.showShieldGlint = b, SHIELD_GLINT_TOOLTIP));
+
+        children.addAll(buttons);
     }
 
     protected int getWindowLeft() {
@@ -102,40 +97,13 @@ public class ArmorConfigWindow extends AbstractParentElement implements Drawable
         return 10;
     }
 
-    protected SliderWidget getTransparencySlider(EquipmentSlot slot, int x, int y, String translationKey) {
-        var initialValue = armorConfig.getTransparency(slot);
-
-        return new SliderWidget(getWindowLeft() + x, getWindowTop() + y,
-                77, 20, Text.translatable(translationKey, initialValue), initialValue / 100f) {
-            @Override
-            protected void updateMessage() {
-                setMessage(Text.translatable(translationKey, (byte) (this.value * 100)));
-            }
-
-            @Override
-            protected void applyValue() {
-                armorConfig.setTransparency(slot, (byte) (this.value * 100));
-            }
-
-            @Override
-            public void playDownSound(SoundManager soundManager) {
-                soundManager.play(PositionedSoundInstance.master(SoundEvents.UI_BUTTON_CLICK, 1.0F));
-            }
-
-            @Override
-            public void onRelease(double mouseX, double mouseY) {
-            }
-        };
-    }
-
-    protected TexturedButtonWidget getGlintButton(EquipmentSlot slot, int x, int y) {
-        return new ToggleButtonWidget(parent, getWindowLeft() + x, getWindowTop() + y, 0, 38,
-                TEXTURE, armorConfig.getGlint(slot), b -> armorConfig.setGlint(slot, b), GLINT_TOOLTIP);
-    }
-
     @Override
     public void render(MatrixStack matrices, int mouseX, int mouseY, float delta) {
-        renderBackground(matrices, TEXTURE, -999);
+        for (var sliderSetTabs : sliderSetTabs) {
+            sliderSetTabs.render(matrices, mouseX, mouseY, sliderSetTabs == selectedSliderSetTab);
+        }
+
+        renderBackground(matrices, BACKGROUND_TEXTURE, -999);
         for (var drawable : buttons) {
             drawable.render(
                     matrices,
@@ -144,6 +112,12 @@ public class ArmorConfigWindow extends AbstractParentElement implements Drawable
                     delta
             );
         }
+        selectedSliderSetTab.sliderSet.render(
+                matrices,
+                isEditable() ? mouseX : -1,
+                isEditable() ? mouseY : -1,
+                delta
+        );
 
         var playerX = getWindowRight() - 59;
         var playerY = getWindowTop() + 155;
@@ -154,7 +128,7 @@ public class ArmorConfigWindow extends AbstractParentElement implements Drawable
                 getWindowRight() - 110, getWindowTop() + 10, TEXT_COLOR
         );
 
-        InventoryScreen.drawEntity(playerX, playerY, 70, -mouseX + playerX, -mouseY + playerY - 110, player);
+        InventoryScreen.drawEntity(matrices, playerX, playerY, 70, -mouseX + playerX, -mouseY + playerY - 110, player);
 
         if (!isEditable()) {
             renderBackground(matrices, OVERLAY_TEXTURE, 999);
@@ -164,25 +138,24 @@ public class ArmorConfigWindow extends AbstractParentElement implements Drawable
     public void renderBackground(MatrixStack matrices, Identifier backgroundTexture, int zIndex) {
         int leftSide = getWindowLeft() + 3;
         int topSide = getWindowTop();
-        int iterationsNeeded = getHeightIterations();
 
         RenderSystem.setShaderTexture(0, backgroundTexture);
         RenderSystem.enableBlend();
 
         matrices.push();
         matrices.translate(0, 0, zIndex);
-        drawTexture(matrices, leftSide, topSide, 1, 1, 236, 8);
-        for(int i = 0; i < iterationsNeeded; ++i) {
-            drawTexture(matrices, leftSide, topSide + 8 + 16 * i, 1, 10, 236, 16);
-        }
-        drawTexture(matrices, leftSide, topSide + 8 + 16 * iterationsNeeded, 1, 27, 236, 8);
+        drawTexture(matrices, leftSide, topSide, 1, 1, 236, 254);
         matrices.pop();
 
         RenderSystem.disableBlend();
     }
 
-    private boolean isEditable() {
+    public boolean isEditable() {
         return ModConfig.INSTANCE.globalEnabled;
+    }
+
+    public void selectTab(SliderSetTab sliderSetTab) {
+        selectedSliderSetTab = sliderSetTab;
     }
 
     @Override
@@ -191,12 +164,43 @@ public class ArmorConfigWindow extends AbstractParentElement implements Drawable
 
     @Override
     public List<? extends Element> children() {
-        return buttons;
+        return children;
     }
 
     @Override
     public boolean mouseClicked(double mouseX, double mouseY, int button) {
-        return isEditable() && super.mouseClicked(mouseX, mouseY, button);
+        if (!isEditable()) {
+            return false;
+        }
+
+        for (var sliderSetTab : sliderSetTabs) {
+            if (sliderSetTab.isMouseOver(mouseX, mouseY)) {
+                MinecraftClient.getInstance().getSoundManager()
+                        .play(PositionedSoundInstance.master(SoundEvents.UI_BUTTON_CLICK, 1.0F));
+
+                selectTab(sliderSetTab);
+                return true;
+            }
+        }
+
+        if (selectedSliderSetTab.sliderSet.mouseClicked(mouseX, mouseY, button)) {
+            return true;
+        }
+
+        return super.mouseClicked(mouseX, mouseY, button);
+    }
+
+    @Override
+    public boolean mouseDragged(double mouseX, double mouseY, int button, double deltaX, double deltaY) {
+        if (!isEditable()) {
+            return false;
+        }
+
+        if (selectedSliderSetTab.sliderSet.mouseDragged(mouseX, mouseY, button, deltaX, deltaY)) {
+            return true;
+        }
+
+        return super.mouseDragged(mouseX, mouseY, button, deltaX, deltaY);
     }
 
     @Override