+++ CHANGELOG.md
@@ -1,2 +1,1 @@
-- Updated to 1.19.3
-- Added CICADA integration for inter-mod log conversations
+- Fixed an issue with leather armor not being properly hidden.