+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-- Fixed an issue elytra enchant glint not being properly hidden.
+- Updated to 1.19.4
+- Overhauled armor config ui to support trims and custom glint transparency